Consolidate Options to a single file
The split repository structure required
us to have multiple files for options, one
for feldspar-language and one for feldspar-compiler.

Since the files from feldspar-compiler are imported
there is no need to have multiple files so put all
options in a single file called Options that
resides under Compiler.
