+++ src/Feldspar/Compiler/Frontend/Interactive/Interface.hs
@@ -34,14 +34,13 @@
 
 module Feldspar.Compiler.Frontend.Interactive.Interface where
 
+import Feldspar.Compiler.Options
 import Feldspar.Core.Frontend (Syntactic, reifyFeld)
-import Feldspar.Core.Interpretation (FeldOpts(..), Target(..))
 import Feldspar.Core.UntypedRepresentation (UntypedFeld)
 import Feldspar.Core.Middleend.PassManager
 import Feldspar.Core.Middleend.FromTyped (FrontendPass, frontend)
 import Feldspar.Compiler.Compiler
 import Feldspar.Compiler.Imperative.FromCore
-import Feldspar.Compiler.Backend.C.Options
 import Feldspar.Compiler.Backend.C.Library
 import Feldspar.Compiler.Backend.C.Platforms (availablePlatforms, platformFromName)
 import Feldspar.Compiler.Imperative.Representation (Module(..))
