+++ src/Feldspar/Compiler/Plugin.hs
@@ -59,9 +59,9 @@ import System.IO.Unsafe (unsafePerformIO)
 -- Feldspar specific
 import Feldspar.Runtime
 import Feldspar.Compiler (compile, defaultOptions)
-import Feldspar.Compiler.Backend.C.Options (Options(..))
 import Feldspar.Compiler.Backend.C.Library (encodeFunctionName)
 import Feldspar.Compiler.Marshal ()
+import Feldspar.Compiler.Options (Options(..))
 
 -- | Configurable configuration for the loader.
 feldsparPluginConfigWith :: String -> Options -> Config
